Cleaning and relocate every main file into a core sub folder
