+++ src/bsde_solver/core/optimisation/als.py
@@ -1,6 +1,6 @@
-from bsde_solver.tensor.tensor_train import TensorTrain, left_unfold, right_unfold
-from bsde_solver.tensor.tensor_core import TensorCore
-from bsde_solver.tensor.tensor_network import TensorNetwork
+from bsde_solver.core.tensor.tensor_train import TensorTrain, left_unfold, right_unfold
+from bsde_solver.core.tensor.tensor_core import TensorCore
+from bsde_solver.core.tensor.tensor_network import TensorNetwork
 
 import numpy as np
