+++ src/bsde_solver/core/tensor/tensor_core.py
