+++ src/bsde_solver/core/tensor/tensor_train.py
@@ -4,8 +4,8 @@ import numpy as np
 
 from copy import deepcopy
 
-from bsde_solver.tensor.tensor_core import TensorCore
-from bsde_solver.tensor.tensor_network import TensorNetwork
+from bsde_solver.core.tensor.tensor_core import TensorCore
+from bsde_solver.core.tensor.tensor_network import TensorNetwork
 from bsde_solver.utils import batch_qr
