+++ src/bsde_solver/tests/optimisation/test_als.py
@@ -1,8 +1,8 @@
-from bsde_solver.tensor.tensor_train import BatchTensorTrain, TensorTrain, left_unfold, right_unfold
-from bsde_solver.tensor.tensor_core import TensorCore
-from bsde_solver.tensor.tensor_network import TensorNetwork
+from bsde_solver.core.tensor.tensor_train import BatchTensorTrain, TensorTrain, left_unfold, right_unfold
+from bsde_solver.core.tensor.tensor_core import TensorCore
+from bsde_solver.core.tensor.tensor_network import TensorNetwork
+from bsde_solver.core.optimisation.scalar_als import scalar_ALS, batch_scalar_ALS
 from bsde_solver.utils import batch_qr
-from bsde_solver.optimisation.scalar_als import scalar_ALS, batch_scalar_ALS
 
 import numpy as np
 from time import perf_counter
@@ -10,8 +10,8 @@ from time import perf_counter
 if __name__ == "__main__":
 
     seed = 855
-    degree = 8
-    num_assets = 50
+    degree = 4
+    num_assets = 5
 
     shape = [degree for _ in range(num_assets)]
     dim = 3
@@ -23,7 +23,7 @@ if __name__ == "__main__":
     def poly_derivative(x, degree=10):
         return np.array([i * x ** (i - 1) for i in range(degree)]).T
 
-    n_simulations = 5
+    n_simulations = 10000
     phis, dphis = [], []
 
     np.random.seed(seed)
@@ -41,24 +41,24 @@ if __name__ == "__main__":
 
     #################### Single ALS ####################
 
-    start_time = perf_counter()
-    min_tt = []
-    for i in range(n_simulations):
-        new_tt = scalar_ALS(phis[i], result=b[i], n_iter=5, ranks=ranks)
-        min_tt.append(new_tt)
+    # start_time = perf_counter()
+    # min_tt = []
+    # for i in range(n_simulations):
+    #     new_tt = scalar_ALS(phis[i], result=b[i], n_iter=5, ranks=ranks)
+    #     min_tt.append(new_tt)
 
-    end_time = perf_counter() - start_time
-    print("Time:", end_time)
+    # end_time = perf_counter() - start_time
+    # print("Time:", end_time)
 
-    results = []
-    for i in range(n_simulations):
-        result = TensorNetwork(cores=[min_tt[i]]+phis[i], names=["tt"]+[f"phi_{i+1}" for i in range(num_assets)]).contract().view(np.ndarray).squeeze()
-        print("Reconstruction error:", np.linalg.norm(result - b[i]))
-        # print("Result:", np.round(result, 5), "Expected:", np.round(b[i], 5))
+    # results = []
+    # for i in range(n_simulations):
+    #     result = TensorNetwork(cores=[min_tt[i]]+phis[i], names=["tt"]+[f"phi_{i+1}" for i in range(num_assets)]).contract().view(np.ndarray).squeeze()
+    #     print("Reconstruction error:", np.linalg.norm(result - b[i]))
+    #     # print("Result:", np.round(result, 5), "Expected:", np.round(b[i], 5))
 
-        results.append(float(result))
+    #     results.append(float(result))
 
-    print(results)
+    # print(results)
 
     #################### Batch ALS ####################
 
@@ -82,23 +82,32 @@ if __name__ == "__main__":
 
     b = np.random.rand(n_simulations)*10
 
+    print("Start ALS")
     start_time_batch = perf_counter()
-    batch_als = batch_scalar_ALS(phis, b, n_iter=5, ranks=ranks)
+    batch_als = batch_scalar_ALS(phis, b, n_iter=1, ranks=ranks)
     end_time_batch = perf_counter() - start_time_batch
     print("Time:", end_time_batch)
 
-    unbatch_als = batch_als.unbatch()
 
-    results = []
-    for i in range(n_simulations):
-        sample_phis = [TensorCore(phis[j][i], name=f"phi_{j+1}", indices=(f"m_{j+1}",))
-            for j in range(num_assets)
-        ]
-        result = TensorNetwork(cores=[unbatch_als[i]]+sample_phis, names=["tt"]+[f"phi_{i+1}" for i in range(num_assets)]).contract().view(np.ndarray).squeeze()
-        print("Reconstruction error:", np.linalg.norm(result - b[i]))
-        # print("Result:", np.round(result, 5), "Expected:", np.round(b[i], 5))
-        results.append(float(result))
+    batch_phis = TensorNetwork(cores=phis)
+    result = TensorNetwork(cores=[batch_als, batch_phis], names=["tt", "phi"]).contract(batch=True).view(np.ndarray).squeeze()
 
-    print(results)
+    print("Reconstruction error (batch):", np.linalg.norm(result - b))
+
+    print(result[:10], b[:10])
+
+    # unbatch_als = batch_als.unbatch()
+
+    # results = []
+    # for i in range(n_simulations):
+    #     sample_phis = [TensorCore(phis[j][i], name=f"phi_{j+1}", indices=(f"m_{j+1}",))
+    #         for j in range(num_assets)
+    #     ]
+    #     result = TensorNetwork(cores=[unbatch_als[i]]+sample_phis, names=["tt"]+[f"phi_{i+1}" for i in range(num_assets)]).contract().view(np.ndarray).squeeze()
+    #     print("Reconstruction error:", np.linalg.norm(result - b[i]))
+    #     # print("Result:", np.round(result, 5), "Expected:", np.round(b[i], 5))
+    #     results.append(float(result))
+
+    # print(results)
 
     # print("Speed up factor:", end_time / end_time_batch)
