+++ src/bsde_solver/tests/optimisation/test_mals.py
@@ -2,9 +2,9 @@ from time import perf_counter
 
 import numpy as np
 
-from bsde_solver.tensor.tensor_core import TensorCore
-from bsde_solver.tensor.tensor_network import TensorNetwork
-from bsde_solver.optimisation.mals import scalar_MALS
+from bsde_solver.core.tensor.tensor_core import TensorCore
+from bsde_solver.core.tensor.tensor_network import TensorNetwork
+from bsde_solver.core.optimisation.mals import scalar_MALS
 
 def poly(x, degree=10):
     return np.array([x**i for i in range(degree)]).T
@@ -20,7 +20,7 @@ if __name__ == "__main__":
     num_assets = 8
 
     shape = [degree for _ in range(num_assets)]
-    dim = 2
+    dim = 3
     ranks = (1, ) + (dim,) * (num_assets - 1) + (1, )
 
     n_simulations = 1
@@ -28,7 +28,7 @@ if __name__ == "__main__":
 
     np.random.seed(seed)
     for i in range(n_simulations):
-        x = np.random.rand(num_assets)*5
+        x = np.random.rand(num_assets)*4
 
         phi = [TensorCore(poly(x[i], degree=degree), name=f"phi_{i+1}", indices=(f"m_{i+1}",)) for i in range(num_assets)]
         dphi = [TensorCore(poly_derivative(x[i], degree=degree), name=f"dphi_{i+1}", indices=(f"m_{i+1}",),) for i in range(num_assets)]
@@ -44,7 +44,7 @@ if __name__ == "__main__":
     start_time = perf_counter()
     min_tt = []
     for i in range(n_simulations):
-        new_tt = scalar_MALS(phis[i], result=b[i], n_iter=1, ranks=ranks)
+        new_tt = scalar_MALS(phis[i], result=b[i], n_iter=2, ranks=ranks)
         min_tt.append(new_tt)
 
     end_time = perf_counter() - start_time
